+++ lib/eve_online/esi/universe_ancestries.rb
@@ -16,6 +16,10 @@ module EveOnline
           end
       end
 
+      def etag
+        current_etag
+      end
+
       def scope; end
 
       def url